+++ anvil_consortium_manager/views.py
@@ -244,6 +244,12 @@ class AccountDetail(
         context = super().get_context_data(**kwargs)
         # Add an indicator of whether the account is inactive.
         context["is_inactive"] = self.object.status == models.Account.INACTIVE_STATUS
+        edit_permission_codename = (
+            models.AnVILProjectManagerAccess.EDIT_PERMISSION_CODENAME
+        )
+        context["show_edit_links"] = self.request.user.has_perm(
+            "anvil_consortium_manager." + edit_permission_codename
+        )
         context["show_deactivate_button"] = not context["is_inactive"]
         context["show_reactivate_button"] = context["is_inactive"]
         return context
@@ -701,6 +707,12 @@ class ManagedGroupDetail(auth.AnVILConsortiumManagerViewRequired, DetailView):
         context["parent_table"] = tables.GroupGroupMembershipTable(
             self.object.parent_memberships.all(), exclude="child_group"
         )
+        edit_permission_codename = (
+            models.AnVILProjectManagerAccess.EDIT_PERMISSION_CODENAME
+        )
+        context["show_edit_links"] = self.request.user.has_perm(
+            "anvil_consortium_manager." + edit_permission_codename
+        )
         return context
 
 
@@ -960,6 +972,12 @@ class WorkspaceDetail(auth.AnVILConsortiumManagerViewRequired, DetailView):
         context["authorization_domain_table"] = tables.ManagedGroupTable(
             self.object.authorization_domains.all(),
             exclude=["workspace", "number_groups", "number_accounts"],
+        )
+        edit_permission_codename = (
+            models.AnVILProjectManagerAccess.EDIT_PERMISSION_CODENAME
+        )
+        context["show_edit_links"] = self.request.user.has_perm(
+            "anvil_consortium_manager." + edit_permission_codename
         )
         return context
 
@@ -1388,6 +1406,16 @@ class GroupGroupMembershipDetail(auth.AnVILConsortiumManagerViewRequired, Detail
             )
         return obj
 
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        edit_permission_codename = (
+            models.AnVILProjectManagerAccess.EDIT_PERMISSION_CODENAME
+        )
+        context["show_edit_links"] = self.request.user.has_perm(
+            "anvil_consortium_manager." + edit_permission_codename
+        )
+        return context
+
 
 class GroupGroupMembershipCreate(
     auth.AnVILConsortiumManagerEditRequired, SuccessMessageMixin, CreateView
@@ -1524,6 +1552,16 @@ class GroupAccountMembershipDetail(auth.AnVILConsortiumManagerViewRequired, Deta
                 % {"verbose_name": queryset.model._meta.verbose_name}
             )
         return obj
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        edit_permission_codename = (
+            models.AnVILProjectManagerAccess.EDIT_PERMISSION_CODENAME
+        )
+        context["show_edit_links"] = self.request.user.has_perm(
+            "anvil_consortium_manager." + edit_permission_codename
+        )
+        return context
 
 
 class GroupAccountMembershipCreate(
@@ -1686,6 +1724,16 @@ class WorkspaceGroupAccessDetail(auth.AnVILConsortiumManagerViewRequired, Detail
                 % {"verbose_name": queryset.model._meta.verbose_name}
             )
         return obj
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        edit_permission_codename = (
+            models.AnVILProjectManagerAccess.EDIT_PERMISSION_CODENAME
+        )
+        context["show_edit_links"] = self.request.user.has_perm(
+            "anvil_consortium_manager." + edit_permission_codename
+        )
+        return context
 
 
 class WorkspaceGroupAccessCreate(